Refactored code, created custom page margins and max heights
