+++ ComponentLibrary/Pages/ContentView.swift
@@ -16,6 +16,9 @@ struct ContentView: View {
     @Environment(\.openURL) var openURL
     
     
+      private func colorToken(_ token: ColorToken) -> Color {
+            token.color(brand: brand, colorScheme: colorScheme)
+        }
 
     private let cardItems: [(icon: String, text: String)] = [
            (icon: "checkmark.circle", text: "First claim this offer. Then enroll by reviewing the terms and conditions."),
@@ -47,8 +50,7 @@ struct ContentView: View {
                                 
                                     Text("Link text underlined!")
                                                 .underline()
-                                                .foregroundColor(ColorToken.primaryDarkest.color( brand: brand,
-                                                                                                 colorScheme: colorScheme))
+                                                .foregroundColor(     colorToken(.primaryDarkest))
                                                 .typographyStyle(.p2)
                                 
           
@@ -56,14 +58,13 @@ struct ContentView: View {
                                 Text("Primary button text")
                                     .typographyStyle(.button)
                                     .padding()
-                                    .background(ColorToken.containerFillTertiary1.color( brand: brand,colorScheme: colorScheme))
+                                    .background(colorToken(.containerFillTertiary1))
                                     .cornerRadius(8)
                     
                     
                     
                     Text("Sample Success Text")
-                                    .foregroundColor(ColorToken.greenAccessible.color( brand: brand,
-                                                                                       colorScheme: colorScheme))
+                        .foregroundColor(colorToken(.greenAccessible))
                                     .font(.system(size: 26, weight: .heavy))
                     
                     Button("Show Toast") {
@@ -77,9 +78,8 @@ struct ContentView: View {
                     CardComponent(
                                         title: "What should I expect when I enroll in Home Base Essentials?",
                                         items: cardItems,
-                                       iconColor:ColorToken.iconFeedbackSuccess.color( brand: brand,
-                                                colorScheme: colorScheme),
-                                       backgroundColor:ColorToken.containerFillTertiaryDefault.color( brand: brand,colorScheme: colorScheme)
+                                       iconColor:colorToken(.iconFeedbackSuccess),
+                                       backgroundColor:colorToken(.containerFillTertiaryDefault)
                                     )
 
              
@@ -89,14 +89,12 @@ struct ContentView: View {
                 ToastComponent(
                     message: "Complete your Vivint offer by scheduling your installation.",
                     linkText: Text("Schedule installation") .font(.subheadline)
-                        .foregroundColor(ColorToken.grayscale000.color( brand: brand,
-                                                                        colorScheme: colorScheme)).bold(),
+                        .foregroundColor(colorToken(.grayscale000)).bold(),
                     linkAction: {
                         openWebPage("https://www.vivint.com/")
                     },
                     image: Image("doorbell"),
-                    backgroundColor:ColorToken.grayscale800.color( brand: brand,
-                                                                   colorScheme: colorScheme),
+                    backgroundColor:colorToken(.grayscale800),
                     duration: 60.0,
                     isVisible: $showToast
                 )
